Server/Program/requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
python-ldap==3.3.1
schedule==1.2.1